requirements.txt
blinker==1.9.0
click==8.1.7
coverage==7.6.7
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
pytest-cov==6.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
tomli==2.1.0
typing_extensions==4.12.2
Werkzeug==3.1.3
